progress pxX

+++ srcs/ft_nbtostr.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_nbtostr.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mcherel- <mcherel-@student.42.fr>          +#+  +:+       +#+        */
+/*   By: mcherel- <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/12 08:23:17 by mcherel-          #+#    #+#             */
-/*   Updated: 2022/01/23 16:00:59 by mcherel-         ###   ########.fr       */
+/*   Updated: 2022/01/28 10:52:06 by mcherel-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,14 +14,14 @@
 #include <stddef.h>
 //#include <stdio.h>
 
-char	*ft_nbtostr(char *str, long unsigned int nb, int len, char *base)
+char	*ft_nbtostr(char *str, size_t nb, size_t len, char *base)
 {
 	char c;
-	int i;
-	int div;
-	int res;
+	size_t i;
+	size_t div;
+	size_t res;
 	
-	div = (int)ft_strlen(base);/*10 ou 16*/
+	div = ft_strlen(base);/*10 ou 16*/
 	//printf("div : %d\n", div);
 	//printf("base : %s\n", base);
 	//printf("nb : %d\n", nb);
